pyproject.toml
[tool.poetry]
name = "API_CRUD_AUTH"
version = "0.1.0"
description = ""
authors = ["Nicola Gernone"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "3.14"
psycopg2-binary = "^2.9.4"
black = "^23.3.0"
factory-boy = "^3.2.1"
django-model-utils = "^4.3.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.5"
django-allauth = "^0.54.0"
dj-rest-auth = "^4.0.1"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"

[tool.poetry.dev-dependencies]
coverage = "^6.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

